fix(storybook): consolidate to a single .storybook config dir with env-mode split
Collapses the three-directory split (.storybook-config / .storybook-dev /
.storybook-docs) back to a single .storybook directory. The previous layout
triggered NG0201 dependency-injection errors at runtime because Storybook
and @analogjs/storybook-angular have internal references to the literal
".storybook" path that the -c flag does not fully override, causing
preview.ts to load via two module paths and creating duplicate
InjectionToken instances.

The dev/docs split is now driven by the STORYBOOK_MODE environment
variable inside the single .storybook/main.ts:

- dev  → src/components/** (development harnesses)
- docs → src/stories/**   (consumer-facing demos, published to Pages)
- all  → src/**           (default, shows both)

Adds .storybook/manager.ts with the filter-tag sidebar scope used by the
teqbench.website demos iframe. Package scripts set STORYBOOK_MODE per
invocation; legacy storybook / build-storybook scripts map to :dev.

Fixes the NG0201 errors on https://teqbench.github.io/tbx-mat-banners/.
