--- .storybook-docs/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "types": ["node"],
-    "allowImportingTsExtensions": true
-  },
-  "include": ["../src/stories/**/*.stories.ts", "../src/stories/**/*.stories.mdx", "./main.ts", "./preview.ts", "./manager.ts", "../.storybook-config/**/*.ts"],
-  "exclude": ["../src/**/*.spec.ts"]
-}
