+++ eslint.config.js
@@ -4,7 +4,7 @@ import tsdoc from 'eslint-plugin-tsdoc';
 
 export default tseslint.config(
     {
-        ignores: ['.claude/', '.shared-skills/', '.storybook/', '.storybook-config/', '.storybook-dev/', '.storybook-docs/', 'coverage/', 'dist/', 'node_modules/', 'storybook-static/', 'storybook-dev-static/', 'storybook-docs-static/'],
+        ignores: ['.claude/', '.shared-skills/', '.storybook/', 'coverage/', 'dist/', 'node_modules/', 'storybook-static/', 'storybook-dev-static/', 'storybook-docs-static/'],
     },
     ...tseslint.configs.recommended,
     ...angular.configs.tsRecommended,
