+++ package.json
@@ -23,12 +23,12 @@
     "test:coverage": "vitest run --coverage",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "storybook": "storybook dev -p 6006 -c .storybook-dev",
-    "storybook:dev": "storybook dev -p 6006 -c .storybook-dev",
-    "storybook:docs": "storybook dev -p 6007 -c .storybook-docs",
-    "build-storybook": "storybook build -c .storybook-dev -o storybook-dev-static",
-    "build-storybook:dev": "storybook build -c .storybook-dev -o storybook-dev-static",
-    "build-storybook:docs": "storybook build -c .storybook-docs -o storybook-docs-static",
+    "storybook": "STORYBOOK_MODE=dev storybook dev -p 6006 -c .storybook",
+    "storybook:dev": "STORYBOOK_MODE=dev storybook dev -p 6006 -c .storybook",
+    "storybook:docs": "STORYBOOK_MODE=docs storybook dev -p 6007 -c .storybook",
+    "build-storybook": "STORYBOOK_MODE=dev storybook build -c .storybook -o storybook-dev-static",
+    "build-storybook:dev": "STORYBOOK_MODE=dev storybook build -c .storybook -o storybook-dev-static",
+    "build-storybook:docs": "STORYBOOK_MODE=docs storybook build -c .storybook -o storybook-docs-static",
     "prepare": "husky"
   },
   "peerDependencies": {
